Create TotalHandshakes
This problems asks:

If there are n number of people in a conference, and each person shakes hands with every other person exactly one time, how many handshakes will be performed. 

To acquire this formula, I used the logic of: 
At any number of people, the number of handshakes can only be half of the number of people at most. 

a person cannot shake hands with themself, so we need to take them out of the equation. 

arrived at (1/2 * n) * (n -1)