Create MinumumHeightTriangle
This problem asked:
Given integers  b and a, find the smallest integer h , such that there exists a triangle of height h, b base, having an area of at least  a.

first have to solve for h... (2a)/b = h

this did not work as it returns a float. Converted to int using Math.round()

failed 1/3 of test cases..

Math.ceil() was next logical function to use based on the logic of the problem. 

Math.ceil((2a/b)) gave 100% on all test cases